# Throw error when setting invalid timezone