--- a/salt/modules/timezone.py
+++ b/salt/modules/timezone.py
@@ -5,6 +5,7 @@ Module for managing timezone on POSIX-like systems.
 from __future__ import absolute_import
 
 # Import python libs
+import filecmp
 import os
 import errno
 import logging
@@ -280,34 +281,28 @@ def zone_compare(timezone):
     if 'Solaris' in __grains__['os_family']:
         return timezone == get_zone()
 
-    curtzstring = get_zone()
-    if curtzstring != timezone:
-        return False
-
-    tzfile = '/etc/localtime'
-    zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
-
-    if not os.path.exists(tzfile):
-        return 'Error: {0} does not exist.'.format(tzfile)
-
-    hash_type = __opts__.get('hash_type', 'md5')
-
+    tzfile = _get_etc_localtime_path()
+    zonepath = _get_zone_file(timezone)
     try:
-        usrzone = salt.utils.get_hash(zonepath, hash_type)
-    except IOError as exc:
-        raise SaltInvocationError('Invalid timezone \'{0}\''.format(timezone))
+        return filecmp.cmp(tzfile, zonepath, shallow=False)
+    except OSError as exc:
+        problematic_file = exc.filename
+        if problematic_file == zonepath:
+            raise SaltInvocationError(
+                'Can\'t find a local timezone "{0}"'.format(timezone))
+        elif problematic_file == tzfile:
+            raise CommandExecutionError(
+                'Failed to read {0} to determine current timezone: {1}'
+                .format(tzfile, exc.strerror))
+        raise
 
-    try:
-        etczone = salt.utils.get_hash(tzfile, hash_type)
-    except IOError as exc:
-        raise CommandExecutionError(
-            'Problem reading timezone file {0}: {1}'
-            .format(tzfile, exc.strerror)
-        )
 
-    if usrzone == etczone:
-        return True
-    return False
+def _get_etc_localtime_path():
+    return '/etc/localtime'
+
+
+def _get_zone_file(timezone):
+    return '/usr/share/zoneinfo/{0}'.format(timezone)
 
 
 def get_hwclock():