--- a/tests/unit/modules/timezone_test.py
+++ b/tests/unit/modules/timezone_test.py
@@ -1,0 +1,87 @@
+# -*- coding: utf-8 -*-
+
+# Import Python Libs
+from __future__ import absolute_import
+from tempfile import NamedTemporaryFile
+import os
+
+# Import Salt Testing Libs
+from salt.exceptions import CommandExecutionError, SaltInvocationError
+from salttesting import TestCase, skipIf
+from salttesting.helpers import ensure_in_syspath
+from salttesting.mock import (
+    patch,
+    NO_MOCK,
+    NO_MOCK_REASON
+)
+
+ensure_in_syspath('../../')
+
+# Import Salt Libs
+from salt.modules import timezone
+
+# Globals
+timezone.__salt__ = {}
+timezone.__opts__ = {}
+timezone.__grains__ = {}
+
+GET_ZONE_FILE = 'salt.modules.timezone._get_zone_file'
+GET_ETC_LOCALTIME_PATH = 'salt.modules.timezone._get_etc_localtime_path'
+
+
+@skipIf(NO_MOCK, NO_MOCK_REASON)
+@patch.dict(timezone.__grains__, {'os_family': 'Ubuntu'})
+class TimezoneTestCase(TestCase):
+
+    def setUp(self):
+        self.tempfiles = []
+
+    def tearDown(self):
+        for tempfile in self.tempfiles:
+            try:
+                os.remove(tempfile.name)
+            except OSError:
+                pass
+
+    def test_zone_compare_equal(self):
+        etc_localtime = self.create_tempfile_with_contents('a')
+        zone_path = self.create_tempfile_with_contents('a')
+
+        with patch(GET_ZONE_FILE, lambda p: zone_path.name):
+            with patch(GET_ETC_LOCALTIME_PATH, lambda: etc_localtime.name):
+
+                self.assertTrue(timezone.zone_compare('foo'))
+
+    def test_zone_compare_nonexistent(self):
+        etc_localtime = self.create_tempfile_with_contents('a')
+
+        with patch(GET_ZONE_FILE, lambda p: '/foopath/nonexistent'):
+            with patch(GET_ETC_LOCALTIME_PATH, lambda: etc_localtime.name):
+
+                self.assertRaises(SaltInvocationError, timezone.zone_compare, 'foo')
+
+    def test_zone_compare_unequal(self):
+        etc_localtime = self.create_tempfile_with_contents('a')
+        zone_path = self.create_tempfile_with_contents('b')
+
+        with patch(GET_ZONE_FILE, lambda p: zone_path.name):
+            with patch(GET_ETC_LOCALTIME_PATH, lambda: etc_localtime.name):
+
+                self.assertFalse(timezone.zone_compare('foo'))
+
+    def test_missing_localtime(self):
+        with patch(GET_ZONE_FILE, lambda p: '/nonexisting'):
+            with patch(GET_ETC_LOCALTIME_PATH, lambda: '/also-missing'):
+                self.assertRaises(CommandExecutionError, timezone.zone_compare, 'foo')
+
+    def create_tempfile_with_contents(self, contents):
+        temp = NamedTemporaryFile(delete=False)
+        temp.write(contents)
+        temp.close()
+        self.tempfiles.append(temp)
+        return temp
+
+
+if __name__ == '__main__':
+    from integration import run_tests
+    run_tests(TimezoneTestCase, needs_daemon=False)