### What does this PR do?
Ensures that setting an invalid time zone doesn't silently succeed without doing anything.

This check should be vastly faster and simpler than the previous hash comparison, which would require reading both files fully plus computing their hashes to determine if they're different. I suspect that this pattern is sadly common in the rest of the codebase too, as I couldn't find any previous usage of `filecmp` apart from an integration test, but I didn't look for other places where this could make things faster. Comparing hashes do make sense when stuff is remote, but makes absolutely no sense when both files are local. Also removed a call to `timedatectl`, as I didn't see any reason to keep it there.

### What issues does this PR fix or reference?
#37685 

### Previous Behavior
Silently ignored invalid timezones:
```
$ sudo salt-call state.sls timezone
[ERROR   ] Command 'timedatectl set-timezone foobar' failed with return code: 1
[ERROR   ] output: Failed to set time zone: Invalid time zone 'foobar'
local:
----------
          ID: timezone
    Function: timezone.system
        Name: foobar
      Result: True
     Comment: Set timezone foobar
     Started: 14:56:43.656404
    Duration: 29.321 ms
     Changes:
              ----------
              timezone:
                  foobar

Summary for local
------------
Succeeded: 1 (changed=1)
Failed:    0
------------
Total states run:     1
Total run time:  29.321 ms
```

### New Behavior
```
vagrant@debian:~$ sudo salt-call state.sls timezone
[ERROR   ] Unable to compare desired timezone 'foobar' to system timezone: Can't find a local timezone "foobar"
local:
----------
          ID: timezone
    Function: timezone.system
        Name: foobar
      Result: False
     Comment: Unable to compare desired timezone 'foobar' to system timezone: Can't find a local timezone "foobar"
     Started: 06:21:33.183724
    Duration: 1.143 ms
     Changes:   

Summary for local
------------
Succeeded: 0
Failed:    1
------------
Total states run:     1
Total run time:   1.143 ms
```

### Tests written?
You betcha.

Please review [Salt's Contributing Guide](https://docs.saltstack.com/en/latest/topics/development/contributing.html) for best practices.

Closes #37685.