@cachedout 2016-11-28
Beautiful. This really seems like the right way to go.

Could you clean up these lint errors? Then I'll get this merged this morning. Thanks, @thusoy 

https://jenkins.saltstack.com/job/PR/job/salt-pr-lint-n/7054/violations/file/tests/unit/modules/timezone_test.py/