@thusoy 2016-11-28
@cachedout Certainly! I also added a `tearDown` to explicitly clear out the tempfiles.